remove cuego.Complete from cuecfg.Marshal (#1349)
## Summary

`cuego.Complete` leads to non-trivial complexity when dealing with
custom json
marshalling. In particular, it does a round-trip JSON marshal/unmarshal
and fields
that are not handled by golang's json-framework can be lost. This
interferes
with custom marshalling/unmarshalling.

`cuego.Complete` is only needed if we rely on the `cue:` annotations to
autocomplete some fields (I mean, assign default values respecting cue
constraints). We don't rely on this, so, right now it is superfluous.

I also remove the `cue:` annotations in a couple of structs since we do
not
rely on them.

## How was it tested?

- [x] `devbox init` should produce the expected output
- [x] `devbox add` 
- [x] `devbox rm`

testscript unit tests must pass