introduce testscript-examples updater, and update the examples (#1438)
## Summary

This PR adds `testscripts/testrunner/updater/main.go` that walks the
`examples/`
directory and updates projects that have a `run_test` script defined. 

In the future, perhaps we want this to run on other (non run_test)
projects as
well, but that is left as an exercise for future contributors.

Also includes a fix for `devbox update` for insecure packages.

## How was it tested?

```
export DEVBOX_FEATURE_REMOVE_NIXPKGS=1

devbox run update-examples
```

Note: this adds info to devbox.lock for the Remove Nixpkgs feature that 
is not turned on. However, things should continue to work.