[UX] print statement for recomputing the nixEnv (#1521)
## Summary

This print-dev-env calculation may be a bit slow, so we should print a
statement
for the waiting user.

I do this in `ensurePackagesAreInstalled`, instead of in
`nix.PrintDevEnv` because otherwise we can print it 3x due to our cache
misses.

Note, this PR is focussed on the least-risky change for the release to
improve UX for improved wait times on switching to Remove Nixpkgs mode.

## How was it tested?

compiles

(cherry picked from commit 6b718ca506bfa669079a45201f36048fa0a05d75)