[plugins] Remove trigger package (#1545)
## Summary

Previously, we were expecting `$buildInputs` to match the order of
$PATH, specifically, packages installed first appeared first (in both
$PATH and $buildInputs). This would allow us to prioritize the first
binary that shows up. In a few plugins we add a flake that installs a
wrapped or alternate version of the binary. We were relying on the order
of `$buildInputs` to point to the correct wrapped binary.

It looks like `$buildInputs` is not deterministic, or something in the
mariadb/mysql case caused the order to change.

The solution in this PR is to add a new `__remove_trigger_package` field
that when set to true tells devbox to remove the package that causes the
plugin to be installed. This does not affect plugins installed `include`
field.

This has the additional benefit of not installing unnecessary packages
in nix profile.

## How was it tested?

```bash
devbox init
devbox add mysql
devbox shell
```

Also verified that PHP plugin works as expected. Note that in the PHP
case the user should avoid installing composer themselves. It is
included in our php plugin.