Added json schema for 0.8 version (#1617)
## Summary
Title says it.

I put the file in `.schema/` directory so that in future if we change
the schema we can align it with version. Meaning, we can add new schema
files with version numbers and enforce the schema based on the version
of devbox a user has installed. Putting it in root of this repository
would make the root too crowded especially if we add more schema files.

I'm going to host our schema in this repository and subsequently send a
PR to https://github.com/SchemaStore/schemastore/ with a link to the
schema file in our repo to register our repo in https://schemastore.org

afterwards, I can update the vscode extension to reference the
schemastore.org link and enforce our defined schema on files that are
named `devbox.json`

## How was it tested?
validated schema and examples of devbox.json in
https://www.jsonschemavalidator.net/