[envsec] Add environment flag (#1679)
## Summary

Adds new `--environment` flag that gets passed in to envsec. 

A few notes:

* This flag is present in any command that allows a custom config path.
The idea is that in the future users can add per environment
configuration to their `devbox.json`
* This PR validates in devbox.Open instead of CLI. This can probably be
moved up to CLI code but this would have bloated this PR.
* Related, the way we pass options to devbox.Open is very brittle. It
involves a lot of copy paste that is easy to mess up. We should look to
refactor that, maybe in a follow up.

## How was it tested?

```
devbox run --environment prod echo $SECRET
```