[nix] ensure nix version is >= 2.12 (#1700)
## Summary

Ensures that the nix used is >= 2.12 which is the minimum version we
support.
Our CICD tests run this version, and I believe so does @mohsenari.

I considered renaming the function to `ensureSupportedNixInstalled` but
decided against it. The current name is good enough, given the
function's scope
and location in the package.

## How was it tested?

ran `devbox shell` which results in this function being called.
Toggled the condition to trigger the error to verify it prints as we
want.