[plugins] Fix github canonical name (#1897)
## Summary

Fixes https://github.com/jetpack-io/devbox/issues/1848

github canonical names were `owner-repo` but this breaks when there are
multiple plugins in a single repo and creates possible conflicts with
other plugins.

This renames them to:

`owner.repo.name` 

where `name` is the name in the plugin.json. For backward compatibility
we don't require the name to be present, if it's missing the name is the
directory (with slashes replaced with hyphens)

## How was it tested?

Installed a few plugins from the same repo and inspected virtenv dir.