Build and Install Devbox via Nix Flake (#2308)
## Summary

This adds a flake.nix to our repo for building and distributing Devbox.
This will make it easier for developers who install Devbox with Nix to
use pre-releases, and keep Devbox up-to-date on their system.

This changes our workflow in a few small ways:

1. Whenever `go.mod` changes, we'll need to run `devbox run update-hash`
to update the vendorHash for the flake. You can also run this script
when you run `go mod tidy`
2. Releasing a new version will also require us to update the version
number in the flake.nix. This should be done with a commit ("Bump
version to x.y") that also gets tagged with the release number.
3. For debugging purposes, the flake will add the shortened commit hash
to the version string. This also helps identify if a user built the
binary from a flake.
4. I've added a test to `cli-tests` that will build the flake. I've
confirmed that it will fail if the vendorHash is out of date or
incorrect
5. We'll periodically need to run `nix flake update` to update the
flake.lock file.

## How was it tested?

1. Ran `nix build .` in the Devbox repo, confirmed that the resulting
build worked
2. Tested the Github Action by pushing an invalid vendorHash

---------

Signed-off-by: John Lago <750845+Lagoja@users.noreply.github.com>
Co-authored-by: Greg Curtis <greg.curtis@jetpack.io>