devbox: export API for devbox install (#2437)
Add a top-level package that makes a subset of the `internal/devbox`
functionality (currently just install) public.

There are some places where we want to run devbox commands (such as in
the prefetcher) which requires us to run the CLI externally. Exporting a
package means we can run the command implementation directly and don't
need to install the CLI (similar to what we're doing with autodetect).