[nixpkgs] Use lock.stdenv instead of NixPkgsCommitHash (#2584)
## Summary

`NixPkgsCommitHash` is no longer defaulted so it's best to use
lockfile.Stdenv to get the nixpkgs for patching and telemetry. This
ensures we use the correct resolved version at all times. (also I think,
but have not confirmed that before this change, `newGlibcPatchFlake`
will not work is nixpkgs hash is not specified in devbox.json)

## How was it tested?

unit tests 

## Community Contribution License

All community contributions in this pull request are licensed to the
project
maintainers under the terms of the
[Apache 2 License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request, I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 License as
stated in
the
[Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).