[completion] Fix shell completion when specifying --config flag (#2632)
## Summary

Narrow fix to completion when passing `-c`, `--config` flags.

## How was it tested?

Used tab completion when using `-c` flag and it worked.

## Community Contribution License

All community contributions in this pull request are licensed to the
project
maintainers under the terms of the
[Apache 2 License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request, I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 License as
stated in
the
[Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).